feat(test): @blobscan/api: refactor fixtures + more tests for indexer rute @blobscan/blob-storage-manager: mock BlobStorageManager
