+++ .github/workflows/ci.yml
@@ -48,7 +48,7 @@ jobs:
           version: "1.37.2"
           backend: memory
           scheme: http
-          data: ./packages/test/src/fixtures/storage
+          data: packages/test/src/fixtures/storage
 
       - name: Get pnpm store directory
         id: pnpm-cache
