+++ packages/api/test/helper.ts
@@ -4,6 +4,9 @@ import type {
 } from "@trpc/server/adapters/node-http";
 import jwt from "jsonwebtoken";
 
+import blobStorageManager from "@blobscan/blob-storage-manager/src/__mocks__/BlobStorageManager";
+import { prisma } from "@blobscan/db";
+
 import type { TRPCContext } from "../src/context";
 import { createTRPCContext } from "../src/context";
 import { createTRPCInnerContext } from "../src/context";
@@ -11,81 +14,102 @@ import { appRouter } from "../src/root";
 
 export async function getCaller({
   withClient = false,
-}: { withClient?: boolean } = {}) {
+  mockBlobStorageManager = false,
+}: { withClient?: boolean; mockBlobStorageManager?: boolean } = {}) {
+  let ctx;
   if (withClient) {
     const token = jwt.sign("foobar", "supersecret");
     const req = {
       headers: { authorization: `Bearer ${token}` },
     } as NodeHTTPRequest;
 
-    const ctx = (await createTRPCContext({
+    ctx = (await createTRPCContext({
       req,
       res: {} as NodeHTTPResponse,
     })) as TRPCContext;
-
-    return appRouter.createCaller(ctx);
+  } else {
+    ctx = (await createTRPCInnerContext()) as TRPCContext;
   }
 
-  const ctx = (await createTRPCInnerContext()) as TRPCContext;
+  ctx = mockBlobStorageManager
+    ? { ...ctx, blobStorageManager: blobStorageManager }
+    : ctx;
+
   return appRouter.createCaller(ctx);
 }
 
-export const INDEXER_DATA = {
-  block: {
-    number: 1003,
-    hash: "blockHash003",
-    timestamp: 1630257174,
-    slot: 103,
-    blobGasUsed: "10000",
-    excessBlobGas: "5000",
-  },
-  transactions: [
-    {
-      hash: "txHash008",
-      from: "address9",
-      to: "address3",
-      blockNumber: 1003,
-      gasPrice: "20000",
-      maxFeePerBlobGas: "15000",
-    },
-    {
-      hash: "txHash009",
-      from: "address4",
-      to: "address8",
-      blockNumber: 1003,
-      gasPrice: "10000",
-      maxFeePerBlobGas: "1800",
-    },
-    {
-      hash: "txHash010",
-      from: "address7",
-      to: "address2",
-      blockNumber: 1003,
-      gasPrice: "3000000",
-      maxFeePerBlobGas: "20000",
-    },
-  ],
-  blobs: [
-    {
-      versionedHash: "blobHash008",
-      commitment: "commitment008",
-      data: "0x34567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
-      txHash: "txHash008",
-      index: 0,
-    },
-    {
-      versionedHash: "blobHash009",
-      commitment: "commitment009",
-      data: "0x34567890abcdef1234567890abcdef",
-      txHash: "txHash009",
-      index: 0,
-    },
-    {
-      versionedHash: "blobHash010",
-      commitment: "commitment010",
-      data: "0x34567890abcdef1234567890abcdef1234567890abcdef",
-      txHash: "txHash010",
-      index: 0,
-    },
-  ],
-};
+export async function getIndexedData(caller) {
+  const [block, addresses, txs, blobs, storageRefs, blobsOnTransactions] =
+    await Promise.all([
+      caller.block.getByBlockNumber({
+        number: 1003,
+      }),
+      prisma.address.findMany({
+        select: {
+          address: true,
+        },
+        orderBy: {
+          address: "asc",
+        },
+      }),
+      prisma.transaction.findMany({
+        select: {
+          hash: true,
+          blockNumber: true,
+        },
+        orderBy: {
+          hash: "asc",
+        },
+      }),
+      prisma.blob.findMany({
+        select: {
+          versionedHash: true,
+          commitment: true,
+          size: true,
+          firstBlockNumber: true,
+        },
+        orderBy: {
+          versionedHash: "asc",
+        },
+      }),
+      prisma.blobDataStorageReference.findMany({
+        select: {
+          blobHash: true,
+          blobStorage: true,
+          dataReference: true,
+        },
+        orderBy: {
+          blobHash: "asc",
+        },
+      }),
+      prisma.blobsOnTransactions.findMany({
+        select: {
+          blobHash: true,
+          txHash: true,
+          index: true,
+        },
+        orderBy: {
+          blobHash: "asc",
+        },
+      }),
+    ]);
+
+  return {
+    block,
+    addresses,
+    txs,
+    blobs,
+    storageRefs: storageRefs.sort((a, b) => {
+      // First, compare by blobHash
+      if (a.blobHash < b.blobHash) return -1;
+      if (a.blobHash > b.blobHash) return 1;
+
+      // If blobHash is the same, compare by blobStorage
+      if (a.blobStorage < b.blobStorage) return -1;
+      if (a.blobStorage > b.blobStorage) return 1;
+
+      return 0; // If everything is the same
+    }),
+    blobsOnTransactions,
+  };
+}
