+++ packages/blob-storage-manager/src/__mocks__/BlobStorageManager.ts
@@ -1,0 +1,11 @@
+import { beforeEach } from "vitest";
+import { mockDeep, mockReset } from "vitest-mock-extended";
+
+import type { BlobStorageManager } from "../BlobStorageManager";
+
+beforeEach(() => {
+  mockReset(blobStorageManager);
+});
+
+const blobStorageManager = mockDeep<BlobStorageManager>();
+export default blobStorageManager;
